requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
scikit-learn>=1.0.0
Pillow>=9.0.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
opencv-python>=4.5.0
timm>=0.9.0

# Optional: face detection for preprocessing
# facenet-pytorch>=2.5.0

# Optional: Xception pretrained weights
# pretrainedmodels>=0.7.4